# docs/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bank Cards Management",
    "contact": {
      "name": "Шаманова Елизавета Алексеевна",
      "email": "sandress2000@mail.ru"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/users/admin": {
      "patch": {
        "tags": [
          "User API"
        ],
        "summary": "Запрос на получение роли администратора",
        "description": "Аутентифицированный пользователь может запросить роль ADMIN, передав секрет",
        "operationId": "requestAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Ошибка аутентификации или неверный секрет"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Роль успешно обновлена, возвращен JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JwtResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Получить список пользователей с пагинацией",
        "description": "Доступно только пользователю с ролью ADMIN",
        "operationId": "getAll",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 3
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Отказано в доступе"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT отсутствует или невалидный"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Список пользователей",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "User API"
        ],
        "summary": "Получить информацию текущего пользователя",
        "description": "Возвращает данные пользователя, на основании аутентификации",
        "operationId": "getMe",
        "responses": {
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не аутентифицирован"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "Данные пользователя успешно получены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "JwtResponse": {
        "type": "object",
        "description": "Сгенерированный jwt",
        "properties": {
          "jwt": {
            "type": "string",
            "description": "jwt",
            "example": "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiI5N2UwZjJlNC0wMzM5LTRlODQtOGQ1OS0zOGM1YjY0NjMzMDIiLCJyb2xlIjoiVVNFUiIsImlhdCI6MTc1NjgyMTUyNSwiZXhwIjoxNzU2ODI1MTI1fQ.ImL_47MNFoHI8lUhpWdlhfWJmRlZgu-I7PbudmVcaBrdQOp5GnaCqxkggEO7BRk6ic_Eq3YbOhlNH-YNatLvow"
          }
        }
      },
...
      "AdminRequest": {
        "type": "object",
        "description": "Запрос для получения роли администратора",
        "properties": {
          "secret": {
            "type": "string",
            "description": "Секретный пароль для получения роли администратора. В данный момент он равен d337d192-4421-46fc-b81a-cb3152f3f328",
            "example": "d337d192-4421-46fc-b81a-cb3152f3f328"
          }
        }
      },
...
      "UserInfoResponse": {
        "type": "object",
        "description": "Информация о пользователе",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Уникальный идентификатор пользователя",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "login": {
            "type": "string",
            "description": "Логин пользователя",
            "example": "user123"
          },
          "role": {
            "type": "string",
            "description": "Роль пользователя",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "example": "USER"
          }
        }
      },
...
    }
  }
}